add proxy test, split 3 test scripts
