+++ truffle-config.js
@@ -23,9 +23,9 @@
 //
 // const fs = require('fs');
 // const mnemonic = fs.readFileSync(".secret").toString().trim();
-const PrivateKeyProvider = require("truffle-privatekey-provider");
-const privKeyrinkeby = require("./secret");
-const INFURA_API_KEY = "05aa70b19b7543f5bf120cbeb0a50dda";
+// const PrivateKeyProvider = require("truffle-privatekey-provider");
+// const privKeyrinkeby = require("./secret");
+// const INFURA_API_KEY = "05aa70b19b7543f5bf120cbeb0a50dda";
 
 module.exports = {
   /**
@@ -45,15 +45,15 @@ module.exports = {
     // tab if you use this network and you must also set the `host`, `port` and `network_id`
     // options below to some value.
     //
-    rinkeby: {
-      provider: () => new PrivateKeyProvider(privKeyrinkeby, "https://rinkeby.infura.io/v3/" + INFURA_API_KEY),
-      network_id: '4',
-    },
-    // development: {
-    //  host: "127.0.0.1",     // Localhost (default: none)
-    //  port: 8545,            // Standard Ethereum port (default: none)
-    //  network_id: "*",       // Any network (default: none)
+    // rinkeby: {
+    //   provider: () => new PrivateKeyProvider(privKeyrinkeby, "https://rinkeby.infura.io/v3/" + INFURA_API_KEY),
+    //   network_id: '4',
     // },
+    development: {
+     host: "127.0.0.1",     // Localhost (default: none)
+     port: 7545,            // Standard Ethereum port (default: none)
+     network_id: "5777",       // Any network (default: none)
+    },
     // Another network with more advanced options...
     // advanced: {
     // port: 8777,             // Custom port
@@ -83,7 +83,7 @@ module.exports = {
 
   // Set default mocha options here, use special reporters etc.
   mocha: {
-    // timeout: 100000
+    timeout: 5000
   },
 
   // Configure your compilers
